use unwrapped model for eval, revert to default torch.compile
Eval forward passes use eval_model (unwrapped _orig_mod) to avoid
dynamo recompilation from varying batch sizes and dtypes. Training
forward pass uses the compiled model. Kept the dtype fix in sampling.py
(linspace uses model dtype instead of defaulting to float32).

+++ src/inference/sampling.py
@@ -6,7 +6,7 @@ def sample(v, z, actions, num_steps=10, cfg=1.0, negative_actions=None, cache=No
 
 def sample_with_grad(v, z, actions, num_steps=10, cfg=1.0, negative_actions=None, cache=None):
     device = v.device
-    ts = 1 - t.linspace(0, 1, num_steps+1, device=device)
+    ts = 1 - t.linspace(0, 1, num_steps+1, device=device, dtype=v.dtype)
     ts = 3*ts/(2*ts + 1)
     z_prev = z.clone()
     z_prev = z_prev.to(device)
